feat: reset toggle logic initial commit
